Various bug-fixes and performance improvements. Updated User Guide for a number of new features. Some changes to code wording/style choices. Extended support and flexibility for many modules, in particular for star and planet and galaxy formation modules. Expanded RHD support and more accurate integration options and treatment of relativistic effects in the reduced-speed-of-light limit. Expanded chemistry modules and options. Updated turbulent-driving subroutines for easier targeting of desired parameters. Substantial expansion of hooks between radiation and other modules such as chemistry and dust. Substantial expansion of dust dynamics support. Added support for radiation terms in cosmological expansion. Debugging and improved accuracy of various parameters in sink particle modules. Expansion and more detail for default included code thermo-chemistry.

+++ system/ngb_codeblock_before_condition.h
@@ -7,7 +7,7 @@
   // cache some global vars locally for improved compiler alias analysis
   int maxPart = All.MaxPart;
   int maxNodes = MaxNodes;
-  int bunchSize = All.BunchSize;
+  long bunchSize = All.BunchSize;
   integertime ti_Current = All.Ti_Current;
   MyDouble dx, dy, dz, dist, xtmp; xtmp=0;
